Add province test files (API)

+++ api/app/Http/Controllers/ProvinceController.php
@@ -9,6 +9,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 // Models.
+use App\Facility;
 use App\Province;
 
 // Requests.
@@ -101,6 +102,12 @@ class ProvinceController extends Controller
     public function destroy(ProvinceRequest $request, $id)
     {
         $p = Province::findOrFail($id);
+
+        // Make sure that the province does not have any facilities.
+        if (Facility::where('provinceId', $p->id)->first()) {
+            abort(400);
+        }
+
         $deletedProvince = $p->toArray();
         $p->delete();
         return $this->toCcArray($deletedProvince);
